Parametrized create_image.sh scripts for CI

+++ openvidu-server/docker/openvidu-proxy/create_image.sh
@@ -1,5 +1,10 @@
-cp ../utils/discover_my_public_ip.sh ./discover_my_public_ip.sh
+VERSION=$1
+if [[ ! -z $VERSION ]]; then
+    cp ../utils/discover_my_public_ip.sh ./discover_my_public_ip.sh
 
-docker build -t openvidu/openvidu-nginx .
+    docker build -t openvidu/openvidu-nginx:$VERSION .
 
-rm ./discover_my_public_ip.sh
+    rm ./discover_my_public_ip.sh
+else
+    echo "Error: You need to specify a version as first argument"
+fi
