+++ openvidu-server/docker/openvidu-server/create_image.sh
@@ -1,7 +1,12 @@
-cp ../../target/openvidu-server-*.jar ./openvidu-server.jar
-cp ../utils/discover_my_public_ip.sh ./discover_my_public_ip.sh
+VERSION=$1
+if [[ ! -z $VERSION ]]; then
+    cp ../../target/openvidu-server-*.jar ./openvidu-server.jar
+    cp ../utils/discover_my_public_ip.sh ./discover_my_public_ip.sh
 
-docker build -t openvidu/openvidu-server .
+    docker build -t openvidu/openvidu-server:$VERSION .
 
-rm ./openvidu-server.jar
-rm ./discover_my_public_ip.sh
+    rm ./openvidu-server.jar
+    rm ./discover_my_public_ip.sh
+else 
+    echo "Error: You need to specify a version as first argument"
+fi
